+++ Example/GlobalPayments-iOS-SDK/Extensions/UIViewExtensions.swift
@@ -8,7 +8,6 @@ extension UIView {
 
     func bindFrameToSuperviewBounds() {
         guard let superview = self.superview else {
-            print("Error! `superview` was nil – call `addSubview(view: UIView)` before calling `bindFrameToSuperviewBounds()` to fix this.")
             return
         }